# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TechStock Backend API
  description: API REST para gestion de productos y operaciones del sistema.
  contact:
    name: Equipo TechStock
  license:
    name: Uso interno
  version: v1
servers:
- url: http://localhost:8080
  description: Servidor local
tags:
- name: Productos
  description: Operaciones CRUD de productos
paths:
  /api/products/{id}:
    get:
      tags:
      - Productos
      summary: Obtener producto por id
      description: Devuelve un producto por su identificador
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Producto encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: Producto no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      tags:
      - Productos
      summary: Actualizar producto
      description: Actualiza un producto existente
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
        required: true
      responses:
        "200":
          description: Producto actualizado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Datos invalidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Producto no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
      - Productos
      summary: Eliminar producto
      description: Elimina un producto por su id
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Producto eliminado
        "404":
          description: Producto no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/products:
    get:
      tags:
      - Productos
      summary: Listar productos
      description: Devuelve una lista paginada de productos
      operationId: getAll
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "id,ASC"
          items:
            type: string
      responses:
        "200":
          description: Lista paginada de productos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseProductResponse"
    post:
      tags:
      - Productos
      summary: Crear producto
      description: Crea un nuevo producto
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
        required: true
      responses:
        "201":
          description: Producto creado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Datos invalidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
components:
  schemas:
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 0
        description:
          type: string
          maxLength: 500
          minLength: 0
        price:
          type: number
          minimum: 0.0
        stock:
          type: integer
          format: int32
      required:
      - price
      - stock
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 0
        description:
          type: string
          maxLength: 500
          minLength: 0
        price:
          type: number
          minimum: 0.0
        stock:
          type: integer
          format: int32
      required:
      - price
      - stock
    PageResponseProductResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean