# src/main/java/com/proyecto/redes/backend/products/dto/UpdateProductRequest.java
package com.proyecto.redes.backend.products.dto;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.PositiveOrZero;
import jakarta.validation.constraints.Size;

import java.math.BigDecimal;

public record UpdateProductRequest(
        @NotBlank(message = "El nombre es obligatorio")
        @Size(max = 120, message = "El nombre debe tener como maximo 120 caracteres")
        String name,

        @NotBlank(message = "La descripcion es obligatoria")
        @Size(max = 500, message = "La descripcion debe tener como maximo 500 caracteres")
        String description,

        @NotNull(message = "El precio es obligatorio")
        @DecimalMin(value = "0.0", inclusive = false, message = "El precio debe ser mayor a 0")
        BigDecimal price,

        @NotNull(message = "El stock es obligatorio")
        @PositiveOrZero(message = "El stock debe ser cero o positivo")
        Integer stock
) {
}