# src/main/java/com/proyecto/redes/backend/shared/api/ApiErrorResponse.java
package com.proyecto.redes.backend.shared.api;

import java.time.Instant;
import java.util.List;

public record ApiErrorResponse(
        Instant timestamp,
        int status,
        String error,
        String message,
        List<String> details
) {
}